python/requirements.txt
tensorflow==2.20.0rc0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
tqdm>=4.66.0